modular separation of lib
